# artists/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Artists' Catalog API Docs",
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/artist/{id}/albums": {
            "get": {
                "summary": "Get Albums By Artist",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/artists": {
            "get": {
                "summary": "Get All Artists",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/album/{id}/tracks": {
            "get": {
                "summary": "Get All Album Tracks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/track/{id}/albums": {
            "get": {
                "summary": "Get All Albums containing definite track",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Track ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}